fix(refactor): add connection receiver and sender module
