+++ cda-comm-doip/src/ecu_connection.rs
@@ -560,7 +560,10 @@ async fn create_tls_stream(
 
 // Allow building CDA without TLS support
 #[cfg(all(not(feature = "openssl"), not(feature = "mbedtls")))]
-#[allow(clippy::unused_async)] // async matches the signature of the TLS-enabled variant
+#[allow(
+    clippy::unused_async,
+    reason = "async matches the signature of the TLS-enabled variant"
+)]
 async fn create_tls_stream(
     _stream: tokio::net::TcpStream,
     _socket_config: DoipSocketConfig,
